requirements.txt
-i https://217.182.160.171:9002/bgruser/stable/
bee-meteo==0.0.3
certifi==2018.11.29
chardet==3.0.4
cookies==2.2.1
funcsigs==1.0.2
idna==2.7
mock==2.0.0
numpy==1.15.4
pandas==0.23.4
pbr==5.1.1
pymongo==3.7.2
python-dateutil==2.7.5
python-forecastio==1.4.0
pytz==2018.7
requests==2.20.1
responses==0.10.4
six==1.11.0
urllib3==1.24.1
